crud --carlos

+++ app/Http/Controllers/UsuariosController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Models\Usuario;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Date;
 
 class UsuariosController extends Controller
 {
@@ -17,7 +16,7 @@ class UsuariosController extends Controller
 
     public function create()
     {
-        return view('clientes.create');
+        return view('gerente.agregar');
     }
 
     public function store(Request $request)
@@ -28,7 +27,7 @@ class UsuariosController extends Controller
         $usuario->nombre_usuario = $request->nusuario;
         $usuario->contraseña = $request->pass;
         $usuario->rol = $request->seleccion;
-        $usuario->fecha_nacimiento = date($request->fecha) ;
+        $usuario->fecha_nacimiento = date($request->fecha);
         $usuario->direccion = $request->direccion;
         $usuario->email = $request->correo;
         $usuario->telefono = $request->telefono;
@@ -37,13 +36,15 @@ class UsuariosController extends Controller
         return redirect()->route('usuarios.index')->with('success', 'Cliente creado correctamente.');
     }
 
-    public function edit(Usuario $cliente)
+    public function edit($alguien)
     {
-        return view('clientes.edit', compact('cliente'));
+        $alguien = Usuario::find($alguien);
+        return view('gerente.editar', compact('alguien'));
     }
 
-    public function update(Request $request, Usuario $usuario)
+    public function update(Request $request, $usuario)
     {
+        $usuario = Usuario::find($usuario);
         $usuario->nombre = $request->nombre;
         $usuario->apellido = $request->apellido;
         $usuario->nombre_usuario = $request->nusuario;
@@ -55,13 +56,17 @@ class UsuariosController extends Controller
         $usuario->telefono = $request->telefono;
         $usuario->save();
 
-        return redirect()->route('clientes.index')->with('success', 'Cliente actualizado correctamente.');
+        return redirect()->route('usuarios.index')->with('success', 'Cliente actualizado correctamente.');
     }
 
-    public function destroy(Usuario $usuario)
+    public function destroy($usuario)
     {
-        $usuario->delete();
-
-        return redirect()->route('clientes.index')->with('success', 'Cliente eliminado correctamente.');
+        $usuario = Usuario::find($usuario);
+        if ($usuario) {
+            $usuario->delete();
+            return redirect()->route('usuarios.index')->with('success', 'Usuario eliminado correctamente.');
+        } else {
+            return redirect()->route('usuarios.index')->with('error', 'No se pudo eliminar el usuario.');
+        }
     }
 }
